+++ lib/src/context.dart
@@ -271,12 +271,21 @@ class EntityQueryDependencyBuilder<Q extends EntityQuery> {
 
   final _Builder<Q> _builder;
 
-  EntityQueryDependencyBuilder<I> listItem<I extends EntityQuery>(
+  EntityQueryDependencyBuilder<I> listItemQuery<I extends EntityQuery>(
     ListItemSelector<Q, I> sel,
     int index,
   ) {
     return EntityQueryDependencyBuilder._(
-      _builder.listItem(sel, index, maybe: false),
+      _builder.listItemQuery(sel, index, maybe: false),
+    );
+  }
+
+  EntityQueryDependencyBuilder<I> listItemQueryByKey<I extends EntityQuery>(
+    ListItemSelector<Q, I> sel,
+    String key,
+  ) {
+    return EntityQueryDependencyBuilder._(
+      _builder.listItemQueryByKey(sel, key, maybe: false),
     );
   }
 
@@ -333,8 +342,8 @@ class EntityQueryDependencyBuilder<Q extends EntityQuery> {
     return _builder.maybeRefValAttr<T>(sel, attrName);
   }
 
-  EntityId listItemId(ListSelector<Q> sel, int index) {
-    return _builder.listItemId(sel, index)!;
+  ListItem listItem(ListSelector<Q> sel, int index) {
+    return _builder.listItem(sel, index);
   }
 
   T listItemValueAttr<T>(ListSelector<Q> sel, String attrName, int index) {
@@ -342,10 +351,34 @@ class EntityQueryDependencyBuilder<Q extends EntityQuery> {
   }
 
   int listItemCounterAttr(ListSelector<Q> sel, String attrName, int index) {
-    return _builder.listItemCounterAtt(sel, attrName, index);
+    return _builder.listItemCounterAttr(sel, attrName, index);
   }
 
-  List<EntityId> listItems(ListSelector<Q> sel) {
+  EntityId listItemValue(ListSelector<Q> sel, String key) {
+    return _builder.listItemValue(sel, key);
+  }
+
+  T listItemValueAttrByKey<T>(
+    ListSelector<Q> sel,
+    String attrName,
+    String itemKey,
+  ) {
+    return _builder.listItemValAttrByKey<T>(sel, attrName, itemKey);
+  }
+
+  int listItemCounterAttrByKey(
+    ListSelector<Q> sel,
+    String attrName,
+    String itemKey,
+  ) {
+    return _builder.listItemCounterAttrByKey(sel, attrName, itemKey);
+  }
+
+  bool listContainsKey(ListSelector<Q> sel, String key) {
+    return _builder.listContainsKey(sel, key);
+  }
+
+  List<ListItem> listItems(ListSelector<Q> sel) {
     return _builder.listItems(sel);
   }
 
@@ -476,20 +509,16 @@ class EntityQueryListHandlerBuilder<Q extends EntityQuery> {
   final _Builder<Q> _builder;
   final ListSelector<Q> _sel;
 
-  void onItemAdded(ChangeHandler<ListViewItemAdded> handler) {
-    _builder.addChangeHandler<ListViewItemAdded>(_sel, handler);
+  void onItemAdded(ChangeHandler<ListPageItemAdded> handler) {
+    _builder.addChangeHandler<ListPageItemAdded>(_sel, handler);
   }
 
-  void onItemRemoved(ChangeHandler<ListViewItemRemoved> handler) {
-    _builder.addChangeHandler<ListViewItemRemoved>(_sel, handler);
+  void onItemRemoved(ChangeHandler<ListPageItemRemoved> handler) {
+    _builder.addChangeHandler<ListPageItemRemoved>(_sel, handler);
   }
 
-  void onItemAddedIfAbsent(ChangeHandler<ListViewItemAddedIfAbsent> handler) {
-    _builder.addChangeHandler<ListViewItemAddedIfAbsent>(_sel, handler);
-  }
-
-  void onCleared(ChangeHandler<ListViewCleared> handler) {
-    _builder.addChangeHandler<ListViewCleared>(_sel, handler);
+  void onCleared(ChangeHandler<ListPageCleared> handler) {
+    _builder.addChangeHandler<ListPageCleared>(_sel, handler);
   }
 }
 
@@ -502,12 +531,22 @@ class MaybeEntityQueryDependencyBuilder<Q extends EntityQuery> {
 
   final _Builder<Q> _builder;
 
-  MaybeEntityQueryDependencyBuilder<I> listItem<I extends EntityQuery>(
+  MaybeEntityQueryDependencyBuilder<I> listItemQuery<I extends EntityQuery>(
     ListItemSelector<Q, I> sel,
     int index,
   ) {
     return MaybeEntityQueryDependencyBuilder._(
-      _builder.listItem(sel, index, maybe: true),
+      _builder.listItemQuery(sel, index, maybe: true),
+    );
+  }
+
+  MaybeEntityQueryDependencyBuilder<I>
+  listItemQueryByKey<I extends EntityQuery>(
+    ListItemSelector<Q, I> sel,
+    String key,
+  ) {
+    return MaybeEntityQueryDependencyBuilder._(
+      _builder.listItemQueryByKey(sel, key, maybe: true),
     );
   }
 
@@ -539,8 +578,8 @@ class MaybeEntityQueryDependencyBuilder<Q extends EntityQuery> {
     throw _builder.maybeRefValAttr(sel, attrName);
   }
 
-  EntityId? listItemId(ListSelector<Q> sel, int index) {
-    return _builder.listItemId(sel, index);
+  ListItem? listItem(ListSelector<Q> sel, int index) {
+    return _builder.listItem(sel, index);
   }
 
   int? listLength(ListSelector<Q> sel) {
@@ -586,7 +625,7 @@ class _Builder<Q extends EntityQuery> {
 
   final Type queryType;
 
-  _Builder<I> listItem<I extends EntityQuery>(
+  _Builder<I> listItemQuery<I extends EntityQuery>(
     ListItemSelector<Q, I> sel,
     int index, {
     required bool maybe,
@@ -600,7 +639,42 @@ class _Builder<Q extends EntityQuery> {
       throw FluirError('index $index is out of bounds for ${list.debugId}');
     }
 
-    var itemId = list.items.elementAt(index);
+    var listItem = list.items.elementAt(index);
+    var itemId = listItem.value;
+    newPath = newPath.append(ActorQueryPath.root(itemId));
+
+    return _Builder.child(
+      queryType,
+      newPath,
+      list.itemHost(index),
+      element,
+      context,
+      depend: depend,
+      maybe: maybe,
+    );
+  }
+
+  _Builder<I> listItemQueryByKey<I extends EntityQuery>(
+    ListItemSelector<Q, I> sel,
+    String key, {
+    required bool maybe,
+  }) {
+    var view = sel(host.query as Q);
+    var newPath = path.append(ActorQueryPath.root(view.name));
+
+    var list = host.children[view.name] as ActorListViewHost;
+
+    // Find the index by key
+    final index = list.items.toList().indexWhere((item) => item.key == key);
+
+    if (index == -1) {
+      throw FluirError(
+        'list item with key "$key" not found in ${list.debugId}',
+      );
+    }
+
+    var listItem = list.items.elementAt(index);
+    var itemId = listItem.value;
     newPath = newPath.append(ActorQueryPath.root(itemId));
 
     return _Builder.child(
@@ -786,7 +860,7 @@ class _Builder<Q extends EntityQuery> {
     return child.refId;
   }
 
-  EntityId? listItemId(ListSelector<Q> sel, int index) {
+  ListItem listItem(ListSelector<Q> sel, int index) {
     var view = sel(host.query as Q);
     var newPath = path.append(ActorQueryPath.root(view.name));
 
@@ -809,13 +883,74 @@ class _Builder<Q extends EntityQuery> {
     }
 
     if (index >= child.items.length) {
+      if (maybe) {
+        throw FluirError('index $index is out of bounds for ${child.debugId}');
+      }
       throw FluirError('index $index is out of bounds for ${child.debugId}');
     }
 
     return child.items.elementAt(index);
   }
 
-  List<EntityId> listItems(ListSelector<Q> sel) {
+  EntityId listItemValue(ListSelector<Q> sel, String key) {
+    var view = sel(host.query as Q);
+    var newPath = path.append(ActorQueryPath.root(view.name));
+
+    if (depend) {
+      element.depend(queryType, newPath, context);
+    }
+
+    var child = host.children[view.name];
+
+    if (child == null) {
+      throw FluirError(
+        'list view host for ${view.name} not found in ${host.debugId}',
+      );
+    }
+
+    if (child is! ActorListViewHost) {
+      throw FluirError(
+        'wrong host type found ${child.runtimeType}, expected: ActorListViewHost',
+      );
+    }
+
+    // Find item by key
+    final item = child.items.firstWhere(
+      (item) => item.key == key,
+      orElse: () => throw FluirError(
+        'list item with key "$key" not found in ${child.debugId}',
+      ),
+    );
+
+    return item.value;
+  }
+
+  bool listContainsKey(ListSelector<Q> sel, String key) {
+    var view = sel(host.query as Q);
+    var newPath = path.append(ActorQueryPath.root(view.name));
+
+    if (depend) {
+      element.depend(queryType, newPath, context);
+    }
+
+    var child = host.children[view.name];
+
+    if (child == null) {
+      throw FluirError(
+        'list view host for ${view.name} not found in ${host.debugId}',
+      );
+    }
+
+    if (child is! ActorListViewHost) {
+      throw FluirError(
+        'wrong host type found ${child.runtimeType}, expected: ActorListViewHost',
+      );
+    }
+
+    return child.items.any((item) => item.key == key);
+  }
+
+  List<ListItem> listItems(ListSelector<Q> sel) {
     var view = sel(host.query as Q);
     var newPath = path.append(ActorQueryPath.root(view.name));
 
@@ -948,15 +1083,44 @@ class _Builder<Q extends EntityQuery> {
     return child.valueAttr<T>(attrName, index);
   }
 
-  int listItemCounterAtt(ListSelector<Q> sel, String attrName, int index) {
-    var view = sel(host.query as Q);
-    var newPath = path.append(ActorQueryPath.root(view.name));
+  T listItemValAttrByKey<T>(
+    ListSelector<Q> sel,
+    String attrName,
+    String itemKey,
+  ) {
+    final view = sel(host.query as Q);
+    final newPath = path.append(ActorQueryPath.root(view.name));
 
     if (depend) {
       element.depend(queryType, newPath, context);
     }
 
-    var child = host.children[view.name];
+    final child = host.children[view.name];
+
+    if (child == null) {
+      throw FluirError(
+        'list view host for ${view.name} not found in ${host.debugId}',
+      );
+    }
+
+    if (child is! ActorListViewHost) {
+      throw FluirError(
+        'wrong host type found ${child.runtimeType}, expected: ActorListViewHost',
+      );
+    }
+
+    return child.valueAttrByKey<T>(attrName, itemKey);
+  }
+
+  int listItemCounterAttr(ListSelector<Q> sel, String attrName, int index) {
+    final view = sel(host.query as Q);
+    final newPath = path.append(ActorQueryPath.root(view.name));
+
+    if (depend) {
+      element.depend(queryType, newPath, context);
+    }
+
+    final child = host.children[view.name];
 
     if (child == null) {
       throw FluirError(
@@ -971,6 +1135,35 @@ class _Builder<Q extends EntityQuery> {
     }
 
     return child.counterAttr(attrName, index);
+  }
+
+  int listItemCounterAttrByKey(
+    ListSelector<Q> sel,
+    String attrName,
+    String itemKey,
+  ) {
+    final view = sel(host.query as Q);
+    final newPath = path.append(ActorQueryPath.root(view.name));
+
+    if (depend) {
+      element.depend(queryType, newPath, context);
+    }
+
+    final child = host.children[view.name];
+
+    if (child == null) {
+      throw FluirError(
+        'list view host for ${view.name} not found in ${host.debugId}',
+      );
+    }
+
+    if (child is! ActorListViewHost) {
+      throw FluirError(
+        'wrong host type found ${child.runtimeType}, expected: ActorListViewHost',
+      );
+    }
+
+    return child.counterAttrByKey(attrName, itemKey);
   }
 
   void addChangeHandler<C extends Change>(