+++ test/query/query_host_test.dart
@@ -94,12 +94,12 @@ void main() {
       })
       ..list('list1', listAttrs, '100:0:0:0', (rb) {
         rb
-          ..item('actor3', (rb) {
+          ..item('xid-1', 'actor3', (rb) {
             rb
               ..val('view5', 'value35', '3500:0:0:0')
               ..val('view6', 'value36', '3610:0:0:0');
           })
-          ..item('actor4', (rb) {
+          ..item('xid-2', 'actor4', (rb) {
             rb
               ..val('view5', 'value45', '4500:0:0:0')
               ..val('view6', 'value46', '4610:0:0:0');
@@ -136,7 +136,12 @@ void main() {
     expect(view4.changeId, '40:0:0:0');
 
     var list = host.children['list1'] as ActorListViewHost;
-    expect(list.items, ['actor3', 'actor4']);
+    // Check ListItems instead of raw EntityIds
+    expect(list.items.length, 2);
+    expect(list.items.elementAt(0).key, 'xid-1');
+    expect(list.items.elementAt(0).value, 'actor3');
+    expect(list.items.elementAt(1).key, 'xid-2');
+    expect(list.items.elementAt(1).value, 'actor4');
     expect(list.changeId, '100:0:0:0');
     // expect(list.state, ActorQueryState.loaded);
     expect(list.counterAttr('counter-attr', 0), 33);